dash: fix Cargo format
The ID is not being specified correctly: "Cargo" is the package manager, "Crates" is the software repository. The content ID should be specified using the latter.

e.g., in IPLab issue 21182 cargo/cargo/-/winapi/0.3.9 should be crate/cratesio/-/winapi/0.3.9.

Excerpt from the documenation:

Steps:

Use cargo to generate a dependency list;
Sort the results and remove duplicates;
Remove empty lines;
Remove references to project code;
Map each line to a ClearlyDefined ID; and
Invoke the tool.
Note that "Cargo" is the package manager, but "Crates" is the software repository. The content ID should be specified using the latter (crate/cratesio/...).

The above example skips code from the Eclipse Zenoh project. Anything that is not third-party content can be removed in a similar manner.

Note that, in order to better leverage ClearlyDefined data, the "v" should not be included in the version number. For example, serde_json v1.0.85 becomes crate/cratesio/-/serde_json/1.0.85.

Addresses: #32

Signed-off-by: Dan Calavrezo <dan.calavrezo.ext@qorix.ai>