dash: fix Cargo format (#36)
The ID is not being specified correctly: "Cargo" is the package manager, "Crates" is the software repository. The content ID should be specified using the latter.

e.g., in IPLab issue 21182 cargo/cargo/-/winapi/0.3.9 should be crate/cratesio/-/winapi/0.3.9.

Note that, in order to better leverage ClearlyDefined data, the "v" should not be included in the version number. For example, serde_json v1.0.85 becomes crate/cratesio/-/serde_json/1.0.85.

Addresses: #32

Signed-off-by: Dan Calavrezo <dan.calavrezo.ext@qorix.ai>